+++ compiler/src/prog8/compiler/Compiler.kt
@@ -71,6 +71,7 @@ class CompilerArguments(val filepath: Path,
                         val ignoreFootguns: Boolean,
                         val profilingInstrumentation: Boolean,
                         val nostdlib: Boolean,
+                        val traceImports: Boolean = false,
                         val symbolDefs: Map<String, String>,
                         val sourceDirs: List<String> = emptyList(),
                         val outputDir: Path = Path(""),
@@ -105,7 +106,7 @@ fun compileProgram(args: CompilerArguments): CompilationResult? {
         var symbolTable: SymbolTable? = null
 
         val totalTime = measureTime {
-            val libraryDirs =  if(compTarget.libraryPath!=null) listOf(compTarget.libraryPath.toString()) else emptyList()
+            val libraryDirs = if(!args.nostdlib && compTarget.libraryPath!=null) listOf(compTarget.libraryPath.toString()) else emptyList()
             val (parseresult, parseDuration) = measureTimedValue {
                  parseMainModule(
                     args.filepath,
@@ -115,7 +116,8 @@ fun compileProgram(args: CompilerArguments): CompilationResult? {
                     libraryDirs,
                     args.cwd,
                     args.quietAll,
-                    args.nostdlib
+                    args.nostdlib,
+                    args.traceImports
                 )
             }
 
@@ -413,12 +415,13 @@ fun parseMainModule(filepath: Path,
                     libraryDirs: List<String>,
                     cwd: Path,
                     quiet: Boolean,
-                    nostdlib: Boolean): Triple<Program, CompilationOptions, List<Path>> {
+                    nostdlib: Boolean,
+                    traceImports: Boolean): Triple<Program, CompilationOptions, List<Path>> {
     val bf = BuiltinFunctionsFacade(BuiltinFunctions)
     val program = Program(filepath.nameWithoutExtension, bf, compTarget, compTarget)
     bf.program = program
 
-    val importer = ModuleImporter(program, compTarget.name, errors, sourceDirs, libraryDirs, cwd, quiet, nostdlib)
+    val importer = ModuleImporter(program, compTarget.name, errors, sourceDirs, libraryDirs, cwd, quiet, nostdlib, traceImports)
     val importedModuleResult = importer.importMainModule(filepath)
     importedModuleResult.onErr { throw it }
     errors.report()